bs,dkfba,smndbf,masdnb
